refactor(sources): sources.ts /utils file
