+++ frontend/src/lib/components/sidebar/sidebar-footer.svelte
@@ -36,7 +36,7 @@
             </Button>
         </Popover.Trigger>
         <Popover.Content
-            class="w-[var(--radix-popover-trigger-width)] min-w-72 rounded-lg p-2"
+            class="w-[var(--radix-popover-trigger-width)] {$sidebarExpanded ? 'min-w-72' : 'min-w-48'} rounded-lg p-2"
             side={"top"}
             sideOffset={0}
         >
